Shell Updates
Moved from ZSH to Nushell which seems like a super interesting shell and
added some new CLI tools like Zoxide which is a smarter version of the
"cd" command. Starship is a great shell prompt btw :P Also I thought it
didn't really make sense to keep an option to enable/disable the shell
and the terminal so I removed it.