Nice Updates
Removed SDDM. Just log into the system using the default prompt or
whatever TTY thingy. And it should launch a basic TUI menu (UWSM) to
choose a Wayland compositor that is registered for UWSM. Also updated
the system and cleaned up the code to remove unnecessary ugly
whitespace. Also updated the Hyprland config to integrate even better
with UWSM.