Code Cleanup
Removed unncessary whitespace and made sure that Git is configured using
the proper username and email instead of my personal username and email.
This is set during the install and can be changed later in the flake.
Also enabled support for basically every hardware out there along with
firmware regardless of the license. And set up the OpenTablet driver for
graphics tablets and Solaar to configure Logitech wireless devices (I
have a g705 wireless mouse).