Migrating Back to Firefox
I feel like Floorp has horrible performance for me. Like it becomes
super slow and laggy after a random period of time making it
ridiculously annoying. I'm going to go back to Firefox and see if I have
the same issues. If not, then I guess I'll stick with Firefox for now.
Also got rid of Carapace it isn't that useful for me.