Full Rewrite Again
Yep I rewrote the entire thing from scratch... Again. From a barebones
NixOS installation. Was a very fun process and I did a lot of things
this time that I didn't before. I wouldn't say it's better but I do feel
like it's organized/structured better now. I was planning to release
this slightly later but I'm too excited haha. It's usable now if you
know how to fetch NixDots and install it, shouldn't be too hard but not
beginner friendly imo. Still a lot of work to be done though.