Slight Improvements
Added some showcase images in README, working on adding SwayOSD so
changing the volume or whatever shows a nice popup or whatever it's
called and some work on Hyprland config to be more... Flexible.
Configuring specific stuff only if something is enabled and all that.
Also removed that one line "stylix.targets" which was completely
unnecessary.