Added Disko
Added the Disko NixOS module. Now Disko is in charge of the disks, so
filesystems and such are no longer defined in
`hardware-configuration.nix`. Also fixed some installation instructions
in the docs because it had some er, logical errors. Also updated the
system and readded Firefox accounts because it's actually pretty handy.