Added Installation Script
Added an installation script to make it slightly more easier to install
NixDots. The number of steps for installation is now shorter and it's
easier for the user to follow through. The user must fetch NixDots and
configure the flake to set a proper username and all before running the
install script with the correct arguments. The script will then handle
formatting the disk the user wants, moving NixDots to the new install,
installing NixOS etc. Stuff that the user shouldn't really have to
bother copy-pasting or anything imo.