Added `nwg-displays`
This application allows the user to easily configure their monitors,
this is essential for multi-monitor setups. I'm using this mainly
because now I can change display settings at runtime using a GUI rather
than editing the Hyprland configuration and rebuilding NixOS. This
allows for quick experimentation duh. Also moved some packages that I
felt was general-purpose and didn't belong with a specific WM/DE or a
custom Nix module/file to `home.nix`. And updated the system as always
of course.