More Flexibility
I completely forgot and never even noticed the fact that the timezone,
locale and keyboard layout was hardcoded. Sure the user could change all
these by themselves but to do that, they'd have to go through a hell of
a lot of files and replace a lot of things. Rather than do that, I
created about 4 more variables in the flake so you can change the locale
or timezone from a single file. By default, the locale and stuff is set
for people living in the UK because that's where I currently live now.
Make sure to change these things before installing NixDots obviously.