Moved JS Folder
Um so the built/deployed mdBook site doesn't work as intended. The links
aren't opening in a new tab. I don't know why, so I'm going to see if
moving the "js" folder to the root of "docs" changes anything since
looking in the build folder, the "js" folder is in the root. I don't
think this is gonna work but... Who cares? I'm gonna test it anyways :P