Cleaned Up `README.md`
Removed broken links and also using reference links here. I made sure
that the introduction chapter of mdBook and `README.md` were the same,
like they had the same information and such. I want both of these to be
in sync rather than one having different or contradictory information.
Anyways I synced it to be sure that all the information is updated and
correct.