Updated Workflow
Cleaned up the workflow and added lots more comments to it so it's very
obvious what's going on. I also made sure that the workflow is
automatically run only when the documentation is changed or more
specifically, the `docs` folder is changed in order to preserve the
limited amount of CI/CD minutes. Now the workflow is run only when the
documentation is changed rather than on every single commit/push to the
`main` branch which unncessarily wastes precious, precious time.