Slight Work
Er so I added a cool new CLI application while also removing one that I
found unnecessary. I also did some work on the documentation, moved a
bit of code around and got rid of a variable to reduce the code clutter
a bit. I need to eliminate as many variables as possible and avoid
passing them to too many files in order to reduce clutter and make it
easier to read and follow the code.