Backdated Flake
Went back to a slightly older Flake lockfile because I think some
updates broke Spotify stopping it from launching, it doesn't even report
any errors or anything. I have a feeling that this older lockfile will
fix the issues since Spotify works just fine on this laptop with this
lockfile, should be the same on the desktop as well.