Removed Timezone
From now on the user has to set the timezone manually/imperatively after
install. They have to do this only once with the `timedatectl
set-timezone <timezone>` command. This way we can avoid having to make
use of a `timezone` variable in the flake and it's easier for the user
to change their timezone and all. This step is easy so it shouldn't be a
big deal right?