Remove More Useless Files
I realized I could shove a lot of small modules into the `default.nix`
file instead to not have as many files and directories so that's exactly
what I'm doing. I managed to get rid of a bunch of files and directories
without making the `default.nix` overly convoluted and hard to read.