Extra Gaming
Added the `nix-gaming` repo in the flake for some extra games (Osu!
stable) and to also improve some game perforamnce in Steam along with
low latency in Pipewire. I also cleaned up a lot of other things on the
way and started using Home Manager options to configure services and
stuff rather than install them and autostart manually because that's not
super ideal. I'll work on getting Hyprpolkit to work now.