Using Hyprland Properly
ALL THIS TIME I WASN'T EVEN USING THE HYPRLAND FLAKE PROPERLY. The only
reason I realized now is that while Hyprland 0.50 was already released,
I was still stuck on 0.49 which is the latest on Nixpkgs even on
unstable. And the flake should be giving me the absolute latest version.
Anyways setting `package` and `portalPackage` in the Home Manager module
for Hyprland is necessary. Now both Hyprland's NixOS and Home Manager
module are using the same versions of Hyprland packages.