Added Yazi
Added Yazi which is a modern and fancy TUI file manager written in Rust,
it has file previews and other really cool stuff. But it still remains
very fast and is very efficient, I love that it uses Vim-like
keybindings out of the box I feel right at home. Anyways I also removed
a deprecated module of `nix-gaming` that I was using accidentally in
order to get rid of an evaluation warning with Steam.