chore: separating NixOS and Home Manager modules
A bunch of the Nix modules in `modules/nixos` is actually a mixture of
both NixOS related configuration and Home Manager related configuration.

So I would configure a lot of things with Home Manager and use some
NixOS system specific options when required. For example, the Fish
shell, I need to use NixOS options to make Bash automatically launch
Fish shell if it isn't already running when the console is interactive
or whatever and I also have to enable Fish with the normal NixOS option
to get full vendor autocompletion.

I'll separate the NixOS and Home Manager specific configurations by
having one file in `modules/nixos` just for NixOS related configuration
and another file with the same name in `modules/home` that does all the
Home Manager related configuration.