General Fixes
Fixed the Fastfetch logo setting so the small logo variant is used for
the current distro instead of being hardcoded to use the NixOS small
logo or just the normal big logo. I also moved some things around and
all to make it more convenient for importing/exporting Home Manager
modules and all.