FUTURE: Migrating From Rofi to Walker
Walker is Wayland native meaning it was built with Wayland in mind and
also it's highly customizable and very powerful. It can do everything
that Rofi can plus more and of course it also has plugin support just
like Rofi. This is literally just a much better and modern alternative
to Rofi for me. It also uses GTK4 which works great on Wayland and is
significantly better on Wayland compared to GTK3.