Better Way to Handle NixDots on Non-NixOS
This way basically merges the inputs of NixDots with the user's inputs
so the user can define their own inputs and Nix will take the inputs of
NixDots and merge it with theirs. This way, they won't have to redefine
any of the inputs in NixDots. The previous approach worked fine too but,
basically it wouldn't let the user define any of their own inputs
because only the inputs of NixDots would be used which is obviously not
ideal.